requirements.txt
akshare>=1.16.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
pyyaml>=6.0
pandas>=1.5.0
numpy>=1.20.0
python-dateutil>=2.8.0
